Fix chainconfig lookup
Bor was trying to look up the chainconfig from the genesis block, but
if the genesis block is loaded from the database the genesis block object
is nil at that point in the code. Fortunately the chainconfig is available
elsewhere without any issue.

Also updates the connect_nodes.py to peer on localhost instead of the network,
as the network ports aren't open from my home network.