Strip unchanged accounts out of StateUpdate call
When the parallel executor runs, fewer accounts are marked as
changed than when the serial executor runs, which seems to come
back to reverts where something got marked as dirty, but changed
back to its original state.

In general it doesn't matter whether these accounts are sent or not,
but we need consistency across nodes, and because some nodes might
complete a block serially while others complete it in parallel,
they produce different account lists, which can cause problems
downstream.