Add destructs check
We were running into issues where blocks processed in parallel were
reporting a different set of destructs than blocks processed sequentially.
Some digging revealed that the destructs that were present on some servers
but not others were cases where ETH was present in an account before a
contract was deployed to that account, which would lead to a "destruct"
indicator at the same time as new account data.

This check normalizes the data reported by the parallel and serial
implementations to the smaller set of data.